move log to lib and schemas to orm
