+++ index.js
@@ -5,45 +5,12 @@ const port = 8010;
 const sqlite3 = require('sqlite3').verbose();
 const db = new sqlite3.Database(':memory:');
 
-const buildSchemas = require('./src/schemas');
-
-// TODO: Dirty Code #1 - Need to find out how to set this cleaner
-const { createLogger, format, transports } = require('winston');
+const buildSchemas = require('./src/orm/schemas');
+const loggerLib = require('./src/lib/logger');
 var util = require('util');
-const fs = require('fs');
-const path = require('path');
 
-const logDir = 'logs';
+const logger = loggerLib();
 
-// Create the log directory if it does not exist
-if (!fs.existsSync(logDir)) {
-    fs.mkdirSync(logDir);
-}
-
-const filename = path.join(logDir, 'error.log');
-
-const logger = createLogger({
-    level: 'info',
-    format: format.combine(
-        format.timestamp({
-        format: 'YYYY-MM-DD HH:mm:ss'
-        }),
-        format.printf(info => `${info.timestamp} ${info.level}: ${info.message}`)
-    ),
-    transports: [
-        new transports.Console({
-        level: 'info',
-        format: format.combine(
-            format.colorize(),
-            format.printf(
-                info => `${info.timestamp} ${info.level}: ${info.message}`
-            )
-        )
-        }),
-        new transports.File({ filename })
-    ]
-    });
-    
 module.exports = {
     middleware: function(req, res, next){
         console.info(req.method, req.url, res.statusCode);
@@ -70,12 +37,11 @@ console.error = function(){
 console.debug = function(){
     logger.debug.apply(logger, formatArgs(arguments));
 };
-// END: Dirty Code #1
 
 db.serialize(() => {
     buildSchemas(db);
 
-    const app = require('./src/app')(db, logger);
+    const app = require('./src/app')(db);
 
     app.listen(port, () => console.log(`App started and listening on port ${port}`));
 });
