+++ tests/rides.test.js
@@ -6,7 +6,7 @@ const sqlite3 = require('sqlite3').verbose();
 const db = new sqlite3.Database(':memory:');
 
 const app = require('../src/app')(db);
-const buildSchemas = require('../src/schemas');
+const buildSchemas = require('../src/orm/schemas');
 
 describe('RIDES API DB connected', () => {
     describe('GET Rides no record(s)', () => {
